Remove duplicate copies of vesPolyDataToTriangleData.{cpp,h}
Now there is only one copy that lives in src/ves.  It is not compiled
into libves because it depends on vtk.  Eventually there could be a
libvesvtk library that mixes ves and vtk.  For now, Xcode projects can
add the source files to be compiled with their app sources.