kiwi: fix bug where inertial motion thread was not stopped
We have to keep the intertial thread from rendering while the
main thread is rendering.  At first we were stopping inertial
motion when other gestures began but this is not enough.

Intertial motion can in some cases be started while other gestures
are already started.  We should fix this eventually but for now
it is just true.  So we need to always stop intertial motion before
scheduling a non-intertial-motion render.