kiwi: change from demand-driven renders to an animation loop
So far we were rendering at every gesture change (or reset camera, etc.).
This can lead to choppy rendering (we are handeling surfaces with
vastly different numbers of vertices/facets).

Now all render calls are to scheduleRender which just asks that a render
be run at the next pass of the animation loop.

We look at the rendering speed and automatically adjust the rendering
speed accordingly.