Finished moving code from iOS KiwiViewer to vesKiwiViewerApp
Now the iOS KiwiViewer app has no dependencies on VES except for the
class vesKiwiViewerApp.  This allows us to refactor the VES api without
breaking KiwiViewer on Android and iOS, as long as vesKiwiViewerApp is
kept up to date.  This also allows us to test the KiwiViewer app core
on desktop operating systems.