kiwiviewer: use touch motion vector for image slice widget
Previously, it just used the Y-delta of the touch event.  Now
it actually uses the motion vector of the touch event and moves
the slice plane in the correct direction.  In a future iteration,
it should actually track the touch location and keep the slice
plane under the touch position.