add point cloud support to kiwi platform and refactor PointCloud demo
The vesKiwiPolyData class now natively handles point cloud datasets.
This commit refactors the PointCloud android demo and the PointCloud
test so that they use the support provided by the kiwi platform,
instead of using a lot of custom code and shaders.

This commit also adds the ROS Turtle dataset as a builtin dataset in
KiwiViewer for iOS and Android.