bug: avoid crash when image widget does not produce a contour
In a future commit I'd like to separate the contour representation
from the vesKiwiImageWidgetRepresentation class.  The image widget
should not manage the contour representation.  But for now, fix
the code so that it will not crash in the case of an image widget
that does not make use of the contour representation.