enh: add asset loading code to PointCloud app so it can load vtp files
Previously, the PointCloud app loaded an ascii vtk file as a string
and used a vtk reader to read from the string.  Now, the app has been
updated to load binary vtp files.  The vtp file is copied from the asset
manager to the sdcard, then the vtk reader loads it from the sdcard.