fix: remove extension check GL call from data loading methods
A new method is added named initGL().  The new method initializes an
instance of vesOpenGLSupport so that the extension can be tested at
any later time without making a GL call.  This allows applications to
test the extension when loading geometry from a non-GL thread.

The Android and iOS versions of KiwiViewer are updated to call the
new method.  The TestKiwiViewer test is also updated.