enh: refactor camera interaction routines
The camera interaction routines are moved to a new class,
vesKiwiCameraInteractor.  The camera rotion inertia code is moved from
the android app into a new class vesKiwiCameraSpinner.  The camera spinner
uses the interactor to move the camera.

The logic for controlling the camera spinner is moved into
vesKiwiViewerApp.  For now, it has to be in vesKiwiViewerApp instead of
the base class vesKiwiBaseApp so that the spinner can play nice with
widgets that are managed at the vesKiwiViewerApp level.  In the future
we should have a more flexible event handling callback system that would
make it easier to have everything work together.