fix: remove assignment of inverseAspect and swap elements
This fixes a bug in vesRenderer::resetCamera(), the camera position was
being computed incorrectly, the camera was always too far away.

The m_aspect variable is used by the vesRenderer::resetCamera() method to
set the camera position w.r.t. focal point so that the scene bounds
fits well in the view.  The resetCamera() method follows the exact
implementation from vtkCamera, which expects m_asepct[0] to contain the
aspect and aspect[1] to contain 1.0.

The inverseAspect() that was previously stored in m_aspect[0] was never
used, except for the incorrect usage in resetCamera().