enh: add vertex opacity uniform
Previsouly, the mapper's solid color was passed into the shader using the
color vertex attribute.  But, if a mapper used per vertex colors instead
of a constant color, then the only way to set opacity would be to update
every element in the vertex color attribute array.  Using a uniform is a
little more flexible now.  A shader can still decide to ignore the
uniform and use the per-vertex opacity.