enh: remove the old Kiwi Xcode project
This is the old version of KiwiViewer, it uses out dated EAGLView, it
will be replaced with the new example projects that use GLKit.

I am leaving the data directory for now, since the data is used by tests
and by the Android app.  In the future, the data directory could be
removed from the source repo and data could be downloaded as a build step.